+++ src/memory.h
@@ -1,11 +1,11 @@
 #ifndef MEMORY_H__
 #define MEMORY_H__
 
-void *xmalloc(size_t size);
+void *xmalloc(u64 size);
 void xfree(void *ptr);
 
-void globalArenaInit(unsigned int capacity);
-void *globalArenaAllocate(unsigned int size);
+void globalArenaInit(u32 capacity);
+void *globalArenaAllocate(u32 size);
 void globalArenaRelease(void);
 void globalArenaPrintStats(void);
